+++ .idea/misc.xml
@@ -52,6 +52,7 @@
         <entry key="app/src/main/res/layout/fragment_setting.xml" value="0.2078125" />
         <entry key="app/src/main/res/layout/fragment_system.xml" value="0.2078125" />
         <entry key="app/src/main/res/layout/fragment_ux.xml" value="0.2890625" />
+        <entry key="app/src/main/res/layout/fragment_ux_button.xml" value="0.2078125" />
         <entry key="app/src/main/res/layout/layout_bottom_sheet.xml" value="0.335" />
         <entry key="app/src/main/res/menu/menu.xml" value="0.28802083333333334" />
         <entry key="app/src/main/res/menu/menu_bottom_appbar.xml" value="0.28802083333333334" />
