+++ app/src/main/java/com/example/materialdesign/view/ux/UXActivity.kt
@@ -1,21 +1,15 @@
 package com.example.materialdesign.view.ux
 
 
-import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
-import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.example.materialdesign.R
-import com.example.materialdesign.databinding.ActivityMainBinding
 import com.example.materialdesign.databinding.ActivityUxBinding
-import com.example.materialdesign.view.MainActivity
-import com.example.materialdesign.view.MainFragment
 
 
 class UXActivity : AppCompatActivity() {
-    lateinit var binding : ActivityUxBinding
+    lateinit var binding: ActivityUxBinding
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,18 +18,21 @@ class UXActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         binding.bottomNavigationViewUX.setOnItemSelectedListener {
-            when(it.itemId){
-                R.id.fragment_ux->{
-                    navigateTo(UXFragment.newInstance())
+            when (it.itemId) {
+                R.id.fragment_ux_text -> {
+                    navigateTo(TextUXFragment.newInstance())
+                }
+                R.id.fragment_ux_button -> {
+                    navigateTo(ButtonUXFragment.newInstance())
                 }
             }
             true
         }
-        binding.bottomNavigationViewUX.selectedItemId = R.id.fragment_ux
+        binding.bottomNavigationViewUX.selectedItemId = R.id.fragment_ux_text
     }
 
-    fun navigateTo(fragment: Fragment){
-        supportFragmentManager.beginTransaction().replace(R.id.container,fragment).commit()
+    fun navigateTo(fragment: Fragment) {
+        supportFragmentManager.beginTransaction().replace(R.id.container, fragment).commit()
     }
 
 }
