+++ app/src/main/res/drawable/ic_baseline_smart_button_24.xml
@@ -1,4 +1,4 @@
-<vector android:height="24dp" android:tint="#FFFFFF"
+<vector android:height="24dp" android:tint="#000000"
     android:viewportHeight="24" android:viewportWidth="24"
     android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="@android:color/white" android:pathData="M22,9v6c0,1.1 -0.9,2 -2,2h-1l0,-2h1V9H4v6h6v2H4c-1.1,0 -2,-0.9 -2,-2V9c0,-1.1 0.9,-2 2,-2h16C21.1,7 22,7.9 22,9zM14.5,19l1.09,-2.41L18,15.5l-2.41,-1.09L14.5,12l-1.09,2.41L11,15.5l2.41,1.09L14.5,19zM17,14l0.62,-1.38L19,12l-1.38,-0.62L17,10l-0.62,1.38L15,12l1.38,0.62L17,14zM14.5,19l1.09,-2.41L18,15.5l-2.41,-1.09L14.5,12l-1.09,2.41L11,15.5l2.41,1.09L14.5,19zM17,14l0.62,-1.38L19,12l-1.38,-0.62L17,10l-0.62,1.38L15,12l1.38,0.62L17,14z"/>
