+++ app/src/main/res/layout/fragment_ux_text.xml
@@ -3,17 +3,22 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/mypalettegreen_100"
     android:gravity="center"
-    android:orientation="vertical"
-    android:background="@color/mypalettegreen_100">
+    android:orientation="vertical">
 
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
         android:background="@color/colorAccent"
-        android:gravity="center">
+        android:gravity="center"
+        android:orientation="vertical">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="@string/notgood" />
 
         <TextView
             android:layout_width="match_parent"
@@ -22,29 +27,31 @@
             android:textAlignment="center"
             android:textColor="@color/black"
             android:textSize="32sp"
-            android:textStyle="bold"/>
+            android:textStyle="bold" />
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="Это основной текст. Он как и текст выше не несет смысловой нагрузки и необходим исключительно для примера."
             android:textAlignment="center"
             android:textColor="@color/black"
-            android:textSize="22sp"/>
+            android:textSize="22sp" />
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="текст сноски"
             android:textAlignment="center"
             android:textColor="@color/color_gray_two"
-            android:textSize="12sp"/>
+            android:textSize="12sp" />
     </LinearLayout>
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
+        android:background="@color/white"
         android:gravity="center"
-        android:background="@color/white">
+        android:orientation="vertical">
 
         <TextView
             android:layout_width="match_parent"
@@ -53,56 +60,66 @@
             android:textAlignment="center"
             android:textColor="@color/textDarkPrimary"
             android:textSize="22sp"
-            android:textStyle="bold"/>
+            android:textStyle="bold" />
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="Это основной текст. Он как и текст выше не несет смысловой нагрузки и необходим исключительно для примера."
             android:textAlignment="center"
             android:textColor="@color/textDarkSecondary"
-            android:textSize="22sp"/>
+            android:textSize="22sp" />
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="текст сноски"
             android:textAlignment="center"
             android:textColor="@color/textDarkDisabled"
-            android:textSize="22sp"/>
+            android:textSize="22sp" />
     </LinearLayout>
+
     <Space
         android:layout_width="match_parent"
-        android:layout_height="50dp"/>
+        android:layout_height="50dp" />
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
         android:background="@color/teal_200"
-        android:gravity="center">
+        android:gravity="center"
+        android:orientation="vertical">
 
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:text="@string/good" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_margin="5dp"
             android:text="Заголовок текста"
             android:textAlignment="center"
             android:textColor="@color/textDarkPrimary"
-            android:textSize="22sp"
-            android:layout_margin="5dp"/>
+            android:textSize="22sp" />
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_margin="5dp"
             android:text="Это основной текст. Он как и текст выше не несет смысловой нагрузки и необходим исключительно для примера."
             android:textAlignment="center"
             android:textColor="@color/textDarkSecondary"
-            android:textSize="22sp"
-            android:layout_margin="5dp"/>
+            android:textSize="22sp" />
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_margin="5dp"
             android:text="текст сноски"
             android:textAlignment="center"
             android:textColor="@color/textDarkDisabled"
-            android:textSize="22sp"
-            android:layout_margin="5dp"/>
+            android:textSize="22sp" />
     </LinearLayout>
 </LinearLayout>
