+++ app/src/main/res/menu/menu_bottom_navigation_view_ux.xml
@@ -1,10 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:id="@+id/fragment_ux"
-        android:icon="@drawable/ic_baseline_add_task_24"
-        android:tooltipText="@string/fragment_ux"
-        android:title="@string/fragment_ux"/>
-
+    <item
+        android:id="@+id/fragment_ux_text"
+        android:icon="@drawable/ic_baseline_font_download_24"
+        android:title="@string/fragment_ux_text"
+        android:tooltipText="@string/fragment_ux_text" />
+    <item
+        android:id="@+id/fragment_ux_button"
+        android:icon="@drawable/ic_baseline_smart_button_24"
+        android:title="@string/fragment_ux_button"
+        android:tooltipText="@string/fragment_ux_button" />
 
 
 </menu>
